move Arch and Mode to mopro-ffi
