+++ cli/Cargo.toml
@@ -19,3 +19,4 @@ zip = "2.2.1"
 reqwest = { version= "0.12.9", features = ["blocking"] }
 indicatif = "0.17.9"
 serde = { version = "1.0.216", features = ["derive"] }
+mopro-ffi = {path = "../mopro-ffi"}
