--- cli/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod constants;
-pub use constants::{AndroidArch, IosArch, Mode};
