+++ cli/src/utils.rs
@@ -1,8 +1,10 @@
 use std::collections::HashMap;
 
+use mopro_ffi::app_config::constants::{AndroidArch, IosArch};
+
 use crate::{
     config::Config,
-    constants::{Adapter, AndroidArch, IosArch, Platform},
+    constants::{Adapter, Platform},
     select::multi_select,
 };
